langchain-lib/pyproject.toml
[project]
name = "langchain-lib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.7",
    "langchain>=0.3.19",
    "python-dotenv>=1.0.1",
    "langchain-google-genai>=2.0.11",
    "langchain-community>=0.3.18",
    "chromadb>=0.6.3",
    "bs4>=0.0.2",
    "wikipedia>=1.4.0",
]
